feat(mail): make the mobile Profile tab a page
The Profile tab opened a bottom sheet while the other three navigated, and the
sheet's Settings row was a pass-through to a page that carried a Profile row of
its own. Profile is a route now and holds the settings list itself, so the tab
lands where that Settings row used to lead — with the tab bar still under it.

The list moves to useSettingsTabs and the pushed sub-page to
MobileSettingsSubPage, both shared with PWASettings, which stays for the sidebar
and the in-thread Block List / Screener links. Log Out asks first: on a page it
is whatever rests at the bottom of a scroll, not the sheet's deliberate two-step.
