+++ frontend/src/apps/mail/components/mobile/MobileTabBar.vue
@@ -2,14 +2,16 @@
 	<!-- Compose FAB — floats above the bar, right thumb zone. Both the FAB and the
 	     bar step aside while a thread is open: the thread's own reply actions own
 	     the bottom edge there (the modals below stay mounted regardless). Hidden in
-	     search results and the screener too — composing isn't part of those tasks. -->
+	     search results, the screener and the profile page too — composing isn't part of
+	     those tasks. -->
 	<Button
 		v-if="
 			!isThreadOpen &&
 			!isMobileSelectionActive &&
 			!isSearchRoute &&
 			!showSearchModal &&
-			!screenerActive
+			!screenerActive &&
+			!profileActive
 		"
 		variant="solid"
 		class="fixed bottom-[calc(5rem+env(safe-area-inset-bottom))] right-4 z-10 !h-14 !w-14 !rounded-full shadow-lg"
@@ -57,11 +59,20 @@
 				<Icon name="search" :class="iconClass(searchActive)" />
 				<span :class="labelClass(searchActive)">{{ __('Search') }}</span>
 			</button>
-			<!-- Profile is a sheet over the current surface (search included), not a navigation —
-			     it must not dismiss the search overlay. -->
-			<button :class="tabClass(isProfileSheetOpen)" @click="openProfileSheet">
-				<Avatar :label="activeAccountName" size="md" class="shrink-0" />
-				<span :class="labelClass(isProfileSheetOpen)">{{ __('Profile') }}</span>
+			<!-- The tab stands for the person, so it carries their photo when there is one
+			     and falls back to the active account's initial. A photo has no stroke to
+			     thicken, so the ring is what stands in for the other tabs' icon: always
+			     drawn, and `ring-current` puts it on the same ink channel they use — it
+			     darkens to 9 with the rest of the tab on selection, at the 1.5px the icons
+			     stroke at. No ring offset: the ring is the avatar's edge, not a halo. -->
+			<button :class="tabClass(profileActive)" @click="openProfile">
+				<Avatar
+					:label="activeAccountName"
+					:image="user.data?.user_image"
+					size="md"
+					class="size-5.5 shrink-0 ring-[1.5px] ring-current"
+				/>
+				<span :class="labelClass(profileActive)">{{ __('Profile') }}</span>
 			</button>
 		</div>
 	</nav>
@@ -69,35 +80,29 @@
 	<SendMail v-model="showSendModal" />
 	<SearchModal v-model="showSearchModal" />
 	<MobileFolderSheet />
-	<MobileProfileSheet />
 </template>
 
 <script setup lang="ts">
-import { computed, ref } from 'vue'
+import { computed, inject, ref } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import { Avatar, Button, FeatherIcon } from 'frappe-ui'
 import { Icon } from 'frappe-ui/icons'
 
 import { getIcon, getMailboxName } from '@/apps/mail/utils'
-import {
-	useFolderSheet,
-	useMobileSelection,
-	useProfileSheet,
-} from '@/apps/mail/utils/composables'
+import { useFolderSheet, useMobileSelection } from '@/apps/mail/utils/composables'
 import { userStore } from '@/apps/mail/stores/user'
 import SearchModal from '@/apps/mail/components/Modals/SearchModal.vue'
 import SendMail from '@/apps/mail/components/SendMail.vue'
 import MobileFolderSheet from '@/apps/mail/components/mobile/MobileFolderSheet.vue'
-import MobileProfileSheet from '@/apps/mail/components/mobile/MobileProfileSheet.vue'
 
 import type { MailboxData } from '@/apps/mail/types'
 
 const route = useRoute()
 const router = useRouter()
 const store = userStore()
+const user = inject('$user') as { data: Record<string, any> }
 const { mailboxes, allInboxesUnread } = store
 const { openFolderSheet } = useFolderSheet()
-const { isProfileSheetOpen, openProfileSheet } = useProfileSheet()
 const { isMobileSelectionActive } = useMobileSelection()
 
 const activeAccountName = computed(
@@ -128,6 +133,7 @@ const mailActive = computed(
 )
 const screenerActive = computed(() => route.name === 'mail-screener')
 const searchActive = computed(() => showSearchModal.value || isSearchRoute.value)
+const profileActive = computed(() => route.name === 'mail-profile')
 
 // The Search tab is a navigation like the others: it lands on the search page (so tab
 // selection stays route-driven — an overlay over a mail route read as two active tabs),
@@ -167,6 +173,14 @@ const openScreener = () => {
 	router.push({ name: 'mail-screener', params: { accountId: store.accountId } })
 }
 
+// Profile is a route now, not a sheet over the current surface — so it dismisses the
+// search overlay on the way, like every other navigating tab.
+const openProfile = () => {
+	showSearchModal.value = false
+	if (profileActive.value) return
+	router.push({ name: 'mail-profile', params: { accountId: store.accountId } })
+}
+
 const screeningEnabled = computed(
 	() =>
 		!!store.userResource?.data?.accounts?.find((a) => a.id === store.accountId)
@@ -203,21 +217,27 @@ const mailUnreadCount = computed(() => {
 // icon's corner rather than sitting on the glyph strokes; bordered to read
 // against the translucent bar.
 const dotClass =
-	'bg-surface-red-6 absolute -right-1.5 -top-1 block size-2.5 rounded-full border border-[var(--surface-base)]'
+	'bg-surface-red-6 absolute -right-1 -top-1 block size-2 rounded-full border border-[var(--surface-base)]'
 
-// Active/inactive contrast rides two channels: ink (9 vs 4 — dropping inactive
-// to 3 read as more disparity but tipped into illegible) and weight (stroke
-// 1.75 vs 1.5, semibold vs medium), so the active tab pops without any label
-// going faint.
+// Active/inactive contrast rides two channels: ink (9 vs 5) and weight (stroke
+// 1.75 vs 1.5, semibold vs medium), so the active tab pops without the rest
+// going faint. Inactive sits at 5, not the 4 used for meta text elsewhere —
+// at 4 the whole bar read as disabled rather than as three tappable tabs.
 const tabClass = (active: boolean) =>
 	[
 		'flex flex-1 flex-col items-center justify-center gap-1',
-		active ? 'text-ink-gray-9' : 'text-ink-gray-4',
+		active ? 'text-ink-gray-9' : 'text-ink-gray-5',
 	].join(' ')
 
 const iconClass = (active: boolean) =>
 	['h-6 w-6 shrink-0', active ? '[stroke-width:1.75]' : '[stroke-width:1.5]'].join(' ')
 
+// 11px sits below the type scale's floor (text-xs is 12), so it's spelled out —
+// along with the 0.02em the scale's own tokens carry, which an arbitrary size
+// doesn't bring with it.
 const labelClass = (active: boolean) =>
-	['text-xs !leading-3', active ? '!font-semibold' : '!font-medium'].join(' ')
+	[
+		'text-[11px] tracking-[0.02em] !leading-3',
+		active ? '!font-semibold' : '!font-medium',
+	].join(' ')
 </script>
